requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.9
django-autoslug==1.9.8
django-debug-toolbar==3.5.0
idna==3.3
pyshorteners==1.0.1
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
